# openapi/risk-guardrail.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Risk Guardrail API",
    "version": "0.1.0",
    "description": "Contract-of-record baseline for the Risk Guardrail System."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "tags": [
    {
      "name": "system",
      "description": "Health and metadata endpoints."
    },
    {
      "name": "ingestion",
      "description": "Canonical event ingestion endpoints."
    },
    {
      "name": "admin",
      "description": "Administrative and reporting endpoints."
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["system"],
        "summary": "Health check",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Service health response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fills:ingest": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Ingest canonical fill events",
        "operationId": "ingestFills",
        "description": "Authenticated ingestion placeholder for canonical fill events.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Ingestion accepted placeholder response"
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Tenant scope violation or insufficient permissions"
          }
        }
      }
    },
    "/v1/accounts:snapshot": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Ingest canonical account snapshots",
        "operationId": "ingestAccountSnapshots",
        "description": "Authenticated account snapshot ingestion with duplicate-safe counting.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Snapshot ingestion accepted with persistence counters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSnapshotIngestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload or snapshot content"
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Tenant scope violation or insufficient permissions"
          }
        }
      }
    },
    "/v1/admin/accounts/{accountId}/unlock": {
      "post": {
        "tags": ["admin"],
        "summary": "Unlock an account",
        "operationId": "unlockAccount",
        "description": "Authenticated admin placeholder for unlock operations.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unlock accepted placeholder response"
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient role or tenant scope violation"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "required": ["status", "service"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ok"]
          },
          "service": {
            "type": "string"
          }
        }
      },
      "AccountSnapshotIngestResponse": {
        "type": "object",
        "required": [
          "status",
          "tenant_id",
          "accepted_count",
          "total_count",
          "persisted_count",
          "duplicate_count"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["accepted"]
          },
          "tenant_id": {
            "type": "string"
          },
          "accepted_count": {
            "type": "integer",
            "minimum": 0
          },
          "total_count": {
            "type": "integer",
            "minimum": 0
          },
          "persisted_count": {
            "type": "integer",
            "minimum": 0
          },
          "duplicate_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      }
    }
  }
}